MUESTRA DETALLE  DETALLEENTRADA

+++ Models/Entrada.cs
@@ -2,8 +2,8 @@
 {
     public class Entrada
     {
-        public int CodigoEntrada { get; set; }
-        public int Correlativo { get; set; }
+        public Int64 CodigoEntrada { get; set; }
+        public Int64 Correlativo { get; set; }
         public DateTime Fecha { get; set; }
         public string? Sucursal { get; set; }
         public string? Descripcion { get; set; }
